Update Dockerfile to work with golang 1.20 base image
The previous base image (Ubuntu 20.10) was a non-LTS release and is
now unsupported so the Docker build command failed because it couldn't
find the apt repos.

Using golang base images simplifies the Dockerfile (since we no longer
need to install golang ourselves) and since they're based on Debian
Bullseye they'll be supported for a while.